test(activemodel): redistribute misc.test.ts blocks (PR 26, partial)
`misc.test.ts` was a 3,036-line catch-all under one `describe("ActiveModel")`.
The audit (PR 26) lays out a deterministic split per inner describe.
This PR moves 8 of the largest, most distinct blocks out — pure test
movement, no rename/reword (per CLAUDE.md).

Moves (line ranges from the original file):
- Dirty Tracking (575-678)            -> dirty.test.ts
- Callbacks (680-783)                 -> callbacks.test.ts
- Callbacks (extended) (1132-1237)    -> callbacks.test.ts
- afterCommit / afterRollback (1514-) -> callbacks.test.ts
- Serialization (784-859)             -> serialization.test.ts
- ComparisonValidator (1238-1389)     -> validations/comparison-validation.test.ts
- UuidType (1394-1454)                -> type/uuid.test.ts (new)
- JsonType (1459-1509)                -> type/json.test.ts (new)

`it` count unchanged (1,535 AM pass), so test:compare layout-matching
gets stricter parity without losing any coverage. Remaining blocks
(Validations sub-suites, Validators (extended), Types, ad-hoc dirty/
errors/etc. blocks) deferred to follow-up PRs.
