+++ packages/activemodel/src/misc.test.ts
@@ -572,291 +572,7 @@ describe("ActiveModel", () => {
   // =========================================================================
   // Phase 1200/1250 — Callbacks and Dirty Tracking
   // =========================================================================
-  describe("Dirty Tracking", () => {
-    class Person extends Model {
-      static {
-        this.attribute("name", "string");
-        this.attribute("age", "integer");
-      }
-    }
 
-    it("not changed initially", () => {
-      const p = new Person({ name: "dean", age: 30 });
-      expect(p.changed).toBe(false);
-      expect(p.changedAttributes).toEqual([]);
-    });
-
-    it("setting attribute will result in change", () => {
-      const p = new Person({ name: "dean" });
-      p.writeAttribute("name", "sam");
-      expect(p.changed).toBe(true);
-      expect(p.changedAttributes).toContain("name");
-    });
-
-    it("attributeWas returns original value", () => {
-      const p = new Person({ name: "dean" });
-      p.writeAttribute("name", "sam");
-      expect(p.attributeWas("name")).toBe("dean");
-    });
-
-    it("changes to attribute values", () => {
-      const p = new Person({ name: "dean" });
-      p.writeAttribute("name", "sam");
-      expect(p.attributeChange("name")).toEqual(["dean", "sam"]);
-    });
-
-    it("list of changed attribute keys", () => {
-      const p = new Person({ name: "dean", age: 30 });
-      p.writeAttribute("name", "sam");
-      p.writeAttribute("age", 31);
-      expect(p.changes).toEqual({
-        name: ["dean", "sam"],
-        age: [30, 31],
-      });
-    });
-
-    it("setting color to same value should not result in change being recorded", () => {
-      const p = new Person({ name: "dean" });
-      p.writeAttribute("name", "dean");
-      expect(p.changed).toBe(false);
-    });
-
-    it("resetting attribute", () => {
-      const p = new Person({ name: "dean" });
-      p.writeAttribute("name", "sam");
-      expect(p.changed).toBe(true);
-      p.writeAttribute("name", "dean");
-      expect(p.changed).toBe(false);
-    });
-
-    it("changing the same attribute multiple times retains the correct original value", () => {
-      const p = new Person({ name: "dean" });
-      p.writeAttribute("name", "sam");
-      p.writeAttribute("name", "bob");
-      expect(p.attributeChange("name")).toEqual(["dean", "bob"]);
-    });
-
-    it("restore_attributes should restore all previous data", () => {
-      const p = new Person({ name: "dean", age: 30 });
-      p.writeAttribute("name", "sam");
-      p.writeAttribute("age", 99);
-      p.restoreAttributes();
-      expect(p.readAttribute("name")).toBe("dean");
-      expect(p.readAttribute("age")).toBe(30);
-      expect(p.changed).toBe(false);
-    });
-
-    it("saving should preserve previous changes", () => {
-      const p = new Person({ name: "dean" });
-      p.writeAttribute("name", "sam");
-      p.changesApplied();
-      expect(p.changed).toBe(false);
-      expect(p.previousChanges).toEqual({ name: ["dean", "sam"] });
-    });
-
-    it("setting new attributes should not affect previous changes", () => {
-      const p = new Person({ name: "dean" });
-      p.writeAttribute("name", "sam");
-      p.changesApplied();
-      p.writeAttribute("name", "bob");
-      expect(p.previousChanges).toEqual({ name: ["dean", "sam"] });
-      expect(p.changes).toEqual({ name: ["sam", "bob"] });
-    });
-
-    it("cast-value-aware: same cast value = no change", () => {
-      class Sized extends Model {
-        static {
-          this.attribute("size", "integer");
-        }
-      }
-      const s = new Sized({ size: "2" }); // cast to 2
-      s.writeAttribute("size", "2.3"); // cast to 2
-      expect(s.changed).toBe(false);
-      s.writeAttribute("size", "5.1"); // cast to 5
-      expect(s.changed).toBe(true);
-    });
-  });
-
-  describe("Callbacks", () => {
-    it("before/after callbacks run in order", () => {
-      const order: string[] = [];
-      class Ordered extends Model {
-        static {
-          this.attribute("name", "string");
-          this.beforeSave(() => {
-            order.push("before");
-          });
-          this.afterSave(() => {
-            order.push("after");
-          });
-        }
-      }
-      const o = new Ordered();
-      o.runCallbacks("save", () => {
-        order.push("action");
-      });
-      expect(order).toEqual(["before", "action", "after"]);
-    });
-
-    it("around callbacks wrap the action", () => {
-      const order: string[] = [];
-      class Around extends Model {
-        static {
-          this.attribute("name", "string");
-          this.aroundSave((_record, proceed) => {
-            order.push("around_before");
-            proceed();
-            order.push("around_after");
-          });
-        }
-      }
-      const a = new Around();
-      a.runCallbacks("save", () => {
-        order.push("action");
-      });
-      expect(order).toEqual(["around_before", "action", "around_after"]);
-    });
-
-    it("further callbacks should not be called if before validation throws abort", () => {
-      const order: string[] = [];
-      class Halting extends Model {
-        static {
-          this.attribute("name", "string");
-          this.beforeSave(() => {
-            order.push("before");
-            return false;
-          });
-          this.afterSave(() => {
-            order.push("after");
-          });
-        }
-      }
-      const h = new Halting();
-      const result = h.runCallbacks("save", () => {
-        order.push("action");
-      });
-      expect(result).toBe(false);
-      expect(order).toEqual(["before"]);
-      expect(order).not.toContain("action");
-      expect(order).not.toContain("after");
-    });
-
-    it("before_validation halting prevents validations from running", () => {
-      class NoValidate extends Model {
-        static {
-          this.attribute("name", "string");
-          this.validates("name", { presence: true });
-          this.beforeValidation(() => false);
-        }
-      }
-      const n = new NoValidate();
-      expect(n.isValid()).toBe(false);
-      expect(n.errors.count).toBe(0); // validations never ran
-    });
-
-    it("complete callback chain", () => {
-      const order: string[] = [];
-      class Full extends Model {
-        static {
-          this.beforeSave(() => {
-            order.push("before_save");
-          });
-          this.aroundSave((_r, proceed) => {
-            order.push("around_before");
-            proceed();
-            order.push("around_after");
-          });
-          this.afterSave(() => {
-            order.push("after_save");
-          });
-        }
-      }
-      new Full().runCallbacks("save", () => {
-        order.push("save");
-      });
-      expect(order).toEqual(["before_save", "around_before", "save", "around_after", "after_save"]);
-    });
-  });
-
-  // =========================================================================
-  // Phase 1300/1350 — Serialization and Naming
-  // =========================================================================
-  describe("Serialization", () => {
-    class Post extends Model {
-      static {
-        this.attribute("title", "string");
-        this.attribute("body", "string");
-        this.attribute("rating", "integer");
-      }
-
-      get summary(): string {
-        return String(this.readAttribute("title")).slice(0, 10);
-      }
-    }
-
-    it("method serializable hash should work", () => {
-      const p = new Post({ title: "Hello", body: "World", rating: 5 });
-      expect(p.serializableHash()).toEqual({
-        title: "Hello",
-        body: "World",
-        rating: 5,
-      });
-    });
-
-    it("method serializable hash should work with only option", () => {
-      const p = new Post({ title: "Hello", body: "World", rating: 5 });
-      expect(p.serializableHash({ only: ["title"] })).toEqual({
-        title: "Hello",
-      });
-    });
-
-    it("method serializable hash should work with except option", () => {
-      const p = new Post({ title: "Hello", body: "World", rating: 5 });
-      expect(p.serializableHash({ except: ["body"] })).toEqual({
-        title: "Hello",
-        rating: 5,
-      });
-    });
-
-    it("method serializable hash should work with methods option", () => {
-      const p = new Post({ title: "Hello World!", body: "c", rating: 3 });
-      const result = p.serializableHash({ methods: ["summary"] });
-      expect(result.summary).toBe("Hello Worl");
-    });
-
-    it("method serializable hash should work with only and methods", () => {
-      const p = new Post({ title: "Test", body: "c", rating: 3 });
-      const result = p.serializableHash({
-        only: ["title"],
-        methods: ["summary"],
-      });
-      expect(Object.keys(result).sort()).toEqual(["summary", "title"]);
-    });
-
-    it("asJson returns same as serializableHash", () => {
-      const p = new Post({ title: "Hello", body: "World", rating: 5 });
-      expect(p.asJson()).toEqual(p.serializableHash());
-    });
-
-    it("toJson returns valid JSON string", () => {
-      const p = new Post({ title: "Hello", body: "World", rating: 5 });
-      const parsed = JSON.parse(p.toJson());
-      expect(parsed.title).toBe("Hello");
-      expect(parsed.rating).toBe(5);
-    });
-
-    it("include as string for single association", () => {
-      const p = new Post({ title: "Hello", body: "World", rating: 5 });
-      const author = { _attributes: new Map([["name", "Alice"]]) };
-      (p as any)._preloadedAssociations = new Map([["author", author]]);
-      const result = p.serializableHash({ include: "author" });
-      expect((result.author as any).name).toBe("Alice");
-    });
-  });
-
-  // =========================================================================
-  // Types — Date, DateTime, Decimal
-  // =========================================================================
   describe("Types", () => {
     describe("DateType", () => {
       const type = new Types.DateType();
@@ -1129,414 +845,6 @@ describe("ActiveModel", () => {
   // =========================================================================
   // Callbacks — after_validation
   // =========================================================================
-  describe("Callbacks (extended)", () => {
-    it("afterValidation runs after validation", () => {
-      const order: string[] = [];
-      class Validated extends Model {
-        static {
-          this.attribute("name", "string");
-          this.validates("name", { presence: true });
-          this.afterValidation(() => {
-            order.push("after_validation");
-          });
-        }
-      }
-      const v = new Validated({ name: "dean" });
-      v.isValid();
-      expect(order).toContain("after_validation");
-    });
-
-    it("afterValidation runs even when invalid", () => {
-      const order: string[] = [];
-      class Validated extends Model {
-        static {
-          this.attribute("name", "string");
-          this.validates("name", { presence: true });
-          this.afterValidation(() => {
-            order.push("after_validation");
-          });
-        }
-      }
-      const v = new Validated();
-      v.isValid();
-      expect(order).toContain("after_validation");
-    });
-
-    it("callback inheritance — child inherits parent callbacks", () => {
-      const order: string[] = [];
-      class Parent extends Model {
-        static {
-          this.attribute("name", "string");
-          this.beforeSave(() => {
-            order.push("parent_before");
-          });
-        }
-      }
-      class Child extends Parent {
-        static {
-          this.beforeSave(() => {
-            order.push("child_before");
-          });
-        }
-      }
-      const c = new Child();
-      c.runCallbacks("save", () => {
-        order.push("action");
-      });
-      expect(order).toContain("parent_before");
-      expect(order).toContain("child_before");
-      expect(order.indexOf("parent_before")).toBeLessThan(order.indexOf("child_before"));
-    });
-
-    it("child callbacks do not affect parent", () => {
-      const parentOrder: string[] = [];
-      const childOrder: string[] = [];
-      class Parent extends Model {
-        static {
-          this.attribute("name", "string");
-          this.beforeSave(() => {
-            parentOrder.push("parent");
-            childOrder.push("parent");
-          });
-        }
-      }
-      class Child extends Parent {
-        static {
-          this.beforeSave(() => {
-            childOrder.push("child");
-          });
-        }
-      }
-      new Parent().runCallbacks("save", () => {
-        parentOrder.push("action");
-      });
-      expect(parentOrder).not.toContain("child");
-    });
-
-    it("custom validate with method name string", () => {
-      class WithMethod extends Model {
-        static {
-          this.attribute("value", "integer");
-          this.validate("validateCustom");
-        }
-        validateCustom() {
-          if (this.readAttribute("value") === 0) {
-            this.errors.add("value", "invalid", { message: "cannot be zero" });
-          }
-        }
-      }
-      expect(new WithMethod({ value: 1 }).isValid()).toBe(true);
-      const w = new WithMethod({ value: 0 });
-      expect(w.isValid()).toBe(false);
-      expect(w.errors.get("value")).toContain("cannot be zero");
-    });
-  });
-
-  // =========================================================================
-  // ComparisonValidator
-  // =========================================================================
-  describe("ComparisonValidator", () => {
-    it("validates greaterThan", () => {
-      class Order extends Model {
-        static {
-          this.attribute("quantity", "integer");
-          this.validates("quantity", { comparison: { greaterThan: 0 } });
-        }
-      }
-      expect(new Order({ quantity: 5 }).isValid()).toBe(true);
-      expect(new Order({ quantity: 0 }).isValid()).toBe(false);
-      expect(new Order({ quantity: -1 }).isValid()).toBe(false);
-    });
-
-    it("validates greaterThanOrEqualTo", () => {
-      class Order extends Model {
-        static {
-          this.attribute("quantity", "integer");
-          this.validates("quantity", { comparison: { greaterThanOrEqualTo: 1 } });
-        }
-      }
-      expect(new Order({ quantity: 1 }).isValid()).toBe(true);
-      expect(new Order({ quantity: 0 }).isValid()).toBe(false);
-    });
-
-    it("validates lessThan", () => {
-      class Rating extends Model {
-        static {
-          this.attribute("score", "integer");
-          this.validates("score", { comparison: { lessThan: 10 } });
-        }
-      }
-      expect(new Rating({ score: 9 }).isValid()).toBe(true);
-      expect(new Rating({ score: 10 }).isValid()).toBe(false);
-    });
-
-    it("validates lessThanOrEqualTo", () => {
-      class Rating extends Model {
-        static {
-          this.attribute("score", "integer");
-          this.validates("score", { comparison: { lessThanOrEqualTo: 10 } });
-        }
-      }
-      expect(new Rating({ score: 10 }).isValid()).toBe(true);
-      expect(new Rating({ score: 11 }).isValid()).toBe(false);
-    });
-
-    it("validates comparison with equal to using numeric", () => {
-      class Confirmation extends Model {
-        static {
-          this.attribute("value", "integer");
-          this.validates("value", { comparison: { equalTo: 42 } });
-        }
-      }
-      expect(new Confirmation({ value: 42 }).isValid()).toBe(true);
-      expect(new Confirmation({ value: 43 }).isValid()).toBe(false);
-    });
-
-    it("validates comparison with other than using numeric", () => {
-      class Item extends Model {
-        static {
-          this.attribute("status", "integer");
-          this.validates("status", { comparison: { otherThan: 0 } });
-        }
-      }
-      expect(new Item({ status: 1 }).isValid()).toBe(true);
-      expect(new Item({ status: 0 }).isValid()).toBe(false);
-    });
-
-    it("validates comparison with proc", () => {
-      class Event extends Model {
-        static {
-          this.attribute("startDate", "date");
-          this.attribute("endDate", "date");
-          this.validates("endDate", {
-            comparison: { greaterThan: (record: any) => record.readAttribute("startDate") },
-          });
-        }
-      }
-      const valid = new Event({
-        startDate: new Date("2024-01-01"),
-        endDate: new Date("2024-01-02"),
-      });
-      expect(valid.isValid()).toBe(true);
-
-      const invalid = new Event({
-        startDate: new Date("2024-01-02"),
-        endDate: new Date("2024-01-01"),
-      });
-      expect(invalid.isValid()).toBe(false);
-    });
-
-    it("validates comparison with greater than using date", () => {
-      const tomorrow = new Date("2024-06-02");
-      class Booking extends Model {
-        static {
-          this.attribute("checkIn", "date");
-          this.validates("checkIn", { comparison: { greaterThanOrEqualTo: tomorrow } });
-        }
-      }
-      expect(new Booking({ checkIn: new Date("2024-06-02") }).isValid()).toBe(true);
-      expect(new Booking({ checkIn: new Date("2024-06-01") }).isValid()).toBe(false);
-    });
-
-    it("validates comparison with greater than using string", () => {
-      class Item extends Model {
-        static {
-          this.attribute("code", "string");
-          this.validates("code", { comparison: { greaterThan: "A" } });
-        }
-      }
-      expect(new Item({ code: "B" }).isValid()).toBe(true);
-      expect(new Item({ code: "A" }).isValid()).toBe(false);
-    });
-
-    it("validates comparison with nil allowed", () => {
-      class Item extends Model {
-        static {
-          this.attribute("quantity", "integer");
-          this.validates("quantity", { comparison: { greaterThan: 0 } });
-        }
-      }
-      expect(new Item({}).isValid()).toBe(true);
-    });
-
-    it("supports custom message", () => {
-      class Item extends Model {
-        static {
-          this.attribute("qty", "integer");
-          this.validates("qty", {
-            comparison: { greaterThan: 0, message: "must be positive" },
-          });
-        }
-      }
-      const item = new Item({ qty: 0 });
-      expect(item.isValid()).toBe(false);
-      expect(item.errors.fullMessages).toContain("Qty must be positive");
-    });
-
-    it("validates comparison of multiple values", () => {
-      class Score extends Model {
-        static {
-          this.attribute("value", "integer");
-          this.validates("value", {
-            comparison: { greaterThanOrEqualTo: 0, lessThanOrEqualTo: 100 },
-          });
-        }
-      }
-      expect(new Score({ value: 50 }).isValid()).toBe(true);
-      expect(new Score({ value: -1 }).isValid()).toBe(false);
-      expect(new Score({ value: 101 }).isValid()).toBe(false);
-    });
-  });
-
-  // =========================================================================
-  // UUID Type
-  // =========================================================================
-  describe("UuidType", () => {
-    it("casts a valid UUID to lowercase", () => {
-      class Item extends Model {
-        static {
-          this.attribute("uuid", "uuid");
-        }
-      }
-      const item = new Item({ uuid: "550E8400-E29B-41D4-A716-446655440000" });
-      expect(item.readAttribute("uuid")).toBe("550e8400-e29b-41d4-a716-446655440000");
-    });
-
-    it("returns null for invalid UUID", () => {
-      class Item extends Model {
-        static {
-          this.attribute("uuid", "uuid");
-        }
-      }
-      const item = new Item({ uuid: "not-a-uuid" });
-      expect(item.readAttribute("uuid")).toBe(null);
-    });
-
-    it("handles null", () => {
-      class Item extends Model {
-        static {
-          this.attribute("uuid", "uuid");
-        }
-      }
-      const item = new Item({});
-      expect(item.readAttribute("uuid")).toBe(null);
-    });
-
-    it("accepts dashless UUID format", () => {
-      class Item extends Model {
-        static {
-          this.attribute("uuid", "uuid");
-        }
-      }
-      const item = new Item({ uuid: "550e8400e29b41d4a716446655440000" });
-      expect(item.readAttribute("uuid")).toBe("550e8400-e29b-41d4-a716-446655440000");
-    });
-
-    it("accepts braced UUID format", () => {
-      class Item extends Model {
-        static {
-          this.attribute("uuid", "uuid");
-        }
-      }
-      const item = new Item({ uuid: "{550e8400-e29b-41d4-a716-446655440000}" });
-      expect(item.readAttribute("uuid")).toBe("550e8400-e29b-41d4-a716-446655440000");
-    });
-
-    it("accepts braced dashless UUID format", () => {
-      class Item extends Model {
-        static {
-          this.attribute("uuid", "uuid");
-        }
-      }
-      const item = new Item({ uuid: "{550E8400E29B41D4A716446655440000}" });
-      expect(item.readAttribute("uuid")).toBe("550e8400-e29b-41d4-a716-446655440000");
-    });
-  });
-
-  // =========================================================================
-  // JSON Type
-  // =========================================================================
-  describe("JsonType", () => {
-    it("casts a JSON string to parsed object", () => {
-      class Config extends Model {
-        static {
-          this.attribute("data", "json");
-        }
-      }
-      const c = new Config({ data: '{"key":"value"}' });
-      expect(c.readAttribute("data")).toEqual({ key: "value" });
-    });
-
-    it("passes through objects", () => {
-      class Config extends Model {
-        static {
-          this.attribute("data", "json");
-        }
-      }
-      const c = new Config({ data: { key: "value" } });
-      expect(c.readAttribute("data")).toEqual({ key: "value" });
-    });
-
-    it("returns null for invalid JSON string", () => {
-      class Config extends Model {
-        static {
-          this.attribute("data", "json");
-        }
-      }
-      const c = new Config({ data: "not json{" });
-      expect(c.readAttribute("data")).toBe(null);
-    });
-
-    it("handles arrays", () => {
-      class Config extends Model {
-        static {
-          this.attribute("tags", "json");
-        }
-      }
-      const c = new Config({ tags: [1, 2, 3] });
-      expect(c.readAttribute("tags")).toEqual([1, 2, 3]);
-    });
-
-    it("handles null", () => {
-      class Config extends Model {
-        static {
-          this.attribute("data", "json");
-        }
-      }
-      const c = new Config({});
-      expect(c.readAttribute("data")).toBe(null);
-    });
-  });
-
-  // =========================================================================
-  // afterCommit / afterRollback callback registration
-  // =========================================================================
-  describe("afterCommit / afterRollback callbacks", () => {
-    it("registers afterCommit callback", () => {
-      class Order extends Model {
-        static {
-          this.attribute("total", "integer");
-          this.afterCommit(() => {});
-        }
-      }
-      // Should not throw
-      expect(new Order({ total: 1 }).isValid()).toBe(true);
-    });
-
-    it("registers afterRollback callback", () => {
-      class Order extends Model {
-        static {
-          this.attribute("total", "integer");
-          this.afterRollback(() => {});
-        }
-      }
-      expect(new Order({ total: 1 }).isValid()).toBe(true);
-    });
-  });
-
-  // =================================================================
-  // Phase: attribute_before_type_cast / hasAttribute
-  // =================================================================
   describe("attributeBeforeTypeCast", () => {
     it("returns the raw value before type casting", () => {
       class Price extends Model {
